Move sync_tasks and helpers from worker.py to tasks.py
Move _format_work_queue, _apply_queue_to_body, _auto_complete_ask_tasks,
sync_tasks, and sync_tasks_background into tasks.py alongside the task
CRUD operations they orchestrate. Worker.py calls them via the tasks
module; events.py and main.py import from kennel.tasks directly. Update
all test imports and patches from kennel.worker to kennel.tasks.
