+++ kennel/worker.py
@@ -178,217 +178,6 @@ def _sanitize_slug(raw: str, fallback: str) -> str:
     return slug
 
 
-def _format_work_queue(task_list: list[dict[str, Any]]) -> str:
-    """Format a task list into work-queue markdown.
-
-    Priority order: CI failures → comment-originated → others.
-    Completed tasks appear in a collapsible ``<details>`` section.
-    Each line includes a ``<!-- type:X -->`` HTML comment for round-tripping.
-    """
-    ci_pending: list[tuple[str, str]] = []
-    comment_pending: list[tuple[str, str]] = []
-    other_pending: list[tuple[str, str]] = []
-    completed: list[tuple[str, str]] = []
-
-    def _fmt(t: dict[str, Any]) -> str:
-        title = t.get("title", "")
-        url = (t.get("thread") or {}).get("url", "")
-        return f"[{title}]({url})" if url else title
-
-    for t in task_list:
-        status = t.get("status", TaskStatus.PENDING)
-        task_type = t.get("type", TaskType.SPEC)
-        display = _fmt(t)
-        if status == TaskStatus.COMPLETED:
-            completed.append((display, task_type))
-        elif status in (TaskStatus.PENDING, TaskStatus.IN_PROGRESS):
-            title = t.get("title", "")
-            if title.startswith("CI failure:"):
-                ci_pending.append((display, task_type))
-            elif t.get("thread"):
-                comment_pending.append((display, task_type))
-            else:
-                other_pending.append((display, task_type))
-
-    pending = ci_pending + comment_pending + other_pending
-    lines: list[str] = []
-    for i, (display, task_type) in enumerate(pending):
-        suffix = " **→ next**" if i == 0 else ""
-        lines.append(f"- [ ] {display}{suffix} <!-- type:{task_type} -->")
-
-    if completed:
-        lines.append("")
-        lines.append(f"<details><summary>Completed ({len(completed)})</summary>")
-        lines.append("")
-        for display, task_type in completed:
-            lines.append(f"- [x] {display} <!-- type:{task_type} -->")
-        lines.append("</details>")
-
-    return "\n".join(lines)
-
-
-def _apply_queue_to_body(body: str, queue: str) -> str:
-    """Replace the WORK_QUEUE_START/END section in a PR body with *queue*.
-
-    Returns *body* unchanged if the markers are absent.
-    """
-    start_marker = "<!-- WORK_QUEUE_START -->"
-    end_marker = "<!-- WORK_QUEUE_END -->"
-    start = body.find(start_marker)
-    end = body.find(end_marker)
-    if start == -1 or end == -1:
-        return body
-    start += len(start_marker)
-    return body[:start] + "\n" + queue + "\n" + body[end:]
-
-
-def _auto_complete_ask_tasks(
-    work_dir: Path,
-    gh: GitHub,
-    repo: str,
-    pr_number: int | str,
-    *,
-    _list_tasks=tasks.list_tasks,
-    _complete_by_id=tasks.complete_by_id,
-) -> None:
-    """Mark pending ASK tasks complete when their review thread is resolved."""
-    task_list = _list_tasks(work_dir)
-    ask_tasks = [
-        t
-        for t in task_list
-        if t.get("status") == TaskStatus.PENDING
-        and t.get("title", "").upper().startswith("ASK:")
-        and t.get("thread")
-    ]
-    if not ask_tasks:
-        return
-
-    try:
-        owner, repo_name = repo.split("/", 1)
-        threads_data = gh.get_review_threads(owner, repo_name, pr_number)
-    except Exception:
-        log.exception("sync_tasks: failed to fetch review threads for ASK resolution")
-        return
-
-    resolved_ids: set[int] = set()
-    for node in (
-        threads_data.get("data", {})
-        .get("repository", {})
-        .get("pullRequest", {})
-        .get("reviewThreads", {})
-        .get("nodes", [])
-    ):
-        if node.get("isResolved"):
-            comments = node.get("comments", {}).get("nodes", [])
-            if comments and comments[0].get("databaseId"):
-                resolved_ids.add(int(comments[0]["databaseId"]))
-
-    for task in ask_tasks:
-        comment_id = (task.get("thread") or {}).get("comment_id")
-        if comment_id and int(comment_id) in resolved_ids:
-            log.info(
-                "sync_tasks: ASK task thread resolved — completing: %s", task["title"]
-            )
-            _complete_by_id(work_dir, task["id"])
-
-
-def sync_tasks(
-    work_dir: Path,
-    gh: GitHub,
-    *,
-    _resolve_git_dir_fn=_resolve_git_dir,
-    _list_tasks=tasks.list_tasks,
-    _auto_complete_ask_tasks_fn=_auto_complete_ask_tasks,
-) -> None:
-    """Sync tasks.json → PR body work queue.
-
-    Python replacement for sync-tasks.sh.  Protected by a flock so concurrent
-    calls silently skip rather than race.  Re-runs if tasks.json changes while
-    the body is being updated.
-    """
-    try:
-        git_dir = _resolve_git_dir_fn(work_dir)
-    except subprocess.CalledProcessError:
-        log.warning("sync_tasks: could not resolve git dir for %s", work_dir)
-        return
-
-    fido_dir = git_dir / "fido"
-    fido_dir.mkdir(parents=True, exist_ok=True)
-    sync_lock_path = fido_dir / "sync.lock"
-    sync_lock_fd = open(sync_lock_path, "w")  # noqa: SIM115
-    try:
-        fcntl.flock(sync_lock_fd, fcntl.LOCK_EX | fcntl.LOCK_NB)
-    except BlockingIOError:
-        log.info("sync_tasks: another sync running — skipping")
-        sync_lock_fd.close()
-        return
-
-    try:
-        state = load_state(fido_dir)
-        issue = state.get("issue")
-        if issue is None:
-            log.info("sync_tasks: no current issue — nothing to sync")
-            return
-
-        try:
-            repo = gh.get_repo_info(cwd=work_dir)
-            user = gh.get_user()
-        except Exception:
-            log.exception("sync_tasks: failed to get repo info or user")
-            return
-
-        pr_data = gh.find_pr(repo, issue, user)
-        if pr_data is None or pr_data.get("state") != "OPEN":
-            log.info("sync_tasks: no open PR for issue #%s — nothing to sync", issue)
-            return
-
-        pr_number = pr_data["number"]
-        _auto_complete_ask_tasks_fn(work_dir, gh, repo, pr_number)
-
-        task_list = _list_tasks(work_dir)
-        if not task_list:
-            log.info("sync_tasks: no tasks — nothing to sync")
-            return
-
-        queue = _format_work_queue(task_list)
-        log.info("sync_tasks: syncing task list → PR #%s", pr_number)
-
-        try:
-            body = gh.get_pr_body(repo, pr_number)
-        except Exception:
-            log.exception("sync_tasks: failed to get PR body")
-            return
-
-        if "WORK_QUEUE_START" not in body:
-            log.info(
-                "sync_tasks: PR #%s has no work queue markers — skipping",
-                pr_number,
-            )
-            return
-
-        new_body = _apply_queue_to_body(body, queue)
-        try:
-            gh.edit_pr_body(repo, pr_number, new_body)
-            log.info("sync_tasks: PR #%s work queue synced", pr_number)
-        except Exception:
-            log.exception("sync_tasks: failed to update PR body")
-    finally:
-        sync_lock_fd.close()
-
-
-def sync_tasks_background(
-    work_dir: Path, gh: GitHub, *, _start=threading.Thread.start
-) -> None:
-    """Launch :func:`sync_tasks` in a daemon background thread."""
-    t = threading.Thread(
-        target=sync_tasks,
-        args=(work_dir, gh),
-        name=f"sync-{work_dir.name}",
-        daemon=True,
-    )
-    _start(t)
-
-
 def claude_start(
     fido_dir: Path,
     model: str = "claude-opus-4-6",
@@ -1055,7 +844,7 @@ class Worker:
         log.info("CI fix done (session=%s)", session_id)
 
         # CI failures have no task entry — no complete call needed
-        sync_tasks(self.work_dir, self.gh)
+        tasks.sync_tasks(self.work_dir, self.gh)
         return True
 
     def _filter_threads(
@@ -1194,7 +983,7 @@ class Worker:
         build_prompt(fido_dir, "comments", context)
         session_id, _ = claude_run(fido_dir, cwd=self.work_dir)
         log.info("threads done (session=%s)", session_id)
-        sync_tasks_background(self.work_dir, self.gh)
+        tasks.sync_tasks_background(self.work_dir, self.gh)
         return True
 
     def ensure_pushed(self, remote: str, slug: str) -> bool | None:
@@ -1288,7 +1077,7 @@ class Worker:
         state.pop("current_task_id", None)
         save_state(fido_dir, state)
         self._abort_task.clear()
-        sync_tasks(self.work_dir, self.gh)
+        tasks.sync_tasks(self.work_dir, self.gh)
 
     def execute_task(
         self,
@@ -1385,7 +1174,7 @@ class Worker:
             state = load_state(fido_dir)
             state.pop("current_task_id", None)
             save_state(fido_dir, state)
-            sync_tasks(self.work_dir, self.gh)
+            tasks.sync_tasks(self.work_dir, self.gh)
         return True
 
     def seed_tasks_from_pr_body(self, repo: str, pr_number: int) -> None:
